refactor

+++ lib/presentation/widgets/expense_widgets/expense_tile.dart
@@ -1,8 +1,8 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:intl/intl.dart';
 import '../../../business_logic/blocs/expense_bloc/expense_bloc.dart';
+import '../../../constants.dart';
 import '../../../data/model/category.dart';
 import '../../../data/model/expense.dart';
 import 'create_new_expense_widget.dart';
@@ -52,17 +52,42 @@ class ExpenseTile extends StatelessWidget {
             },
           ),
           IconButton(
-            icon: new Icon(Icons.delete),
+            icon: const Icon(Icons.delete),
             highlightColor: Colors.grey,
             onPressed: () async {
-              BlocProvider.of<ExpenseBloc>(context)
-                  .add(DeleteExpense(expense, category));
-              BlocProvider.of<ExpenseBloc>(context)
-                  .add(GetExpenses(category, dateTimeRange));
+              _delete(context);
             },
           ),
         ]),
       ),
     );
+  }
+
+  void _delete(BuildContext context) {
+    showDialog(
+        context: context,
+        builder: (BuildContext ctx) {
+          return AlertDialog(
+            title: const Text('Please Confirm'),
+            content: kDeleteConfirmBox,
+            actions: [
+              TextButton(
+                  onPressed: () {
+                    BlocProvider.of<ExpenseBloc>(context)
+                        .add(DeleteExpense(expense, category));
+                    BlocProvider.of<ExpenseBloc>(context)
+                        .add(GetExpenses(category, dateTimeRange));
+                    Navigator.pop(context);
+                  },
+                  child: const Text('Yes')),
+              TextButton(
+                  onPressed: () {
+                    // Close the dialog
+                    Navigator.pop(context);
+                  },
+                  child: const Text('No'))
+            ],
+          );
+        });
   }
 }
